+++ CHANGELOG.md
@@ -10,13 +10,13 @@ All versions prior to 1.0.0 are untracked.
 ## [Unreleased]
 
 - cli: `model_signing sign` now supports the `--oauth-force-oob` option (default: False)
+- Added support for signing and verifying using private Sigstore instances (`--trust_config`).
 
 ## [1.0.1] - 2024-04-18
 
 ### Added
 - Added support for pre v1.0 signatures used in production. This is only provided for verification and replicates the experimental behavior at v0.2, bug for bug.
 - Added support for displaying fingerprints of certificates when using signing certificates
-- Added support for signing and verifying using private Sigstore instances (`--trust_config`).
 
 ### Fixed
 - Fix bug in CLI scripts where even if signature verification failed, the script would also output that verification passed and exit with success error code.